+++ aemedge/blocks/blog-hero/blog-hero.css
@@ -6,14 +6,19 @@ main .blog-hero-container {
   padding: 0;
 }
 
-.hero-container h1 {
+.blog-hero.category {
   text-align: center;
+  padding: 0;
+  min-height: unset;
 }
 
-.blog-hero {
-  position: relative;
-  padding: 32px;
-  min-height: 300px;
+.blog-hero.category.purple {
+  background: rgb(25 23 37);
+  background: linear-gradient(290deg, rgb(25 23 37 / 100%) 25%, rgb(95 43 105 / 100%) 82%);
+}
+
+.hero-container h1 {
+  text-align: center;
 }
 
 .blog-hero h1 {
@@ -21,6 +26,30 @@ main .blog-hero-container {
   margin-left: auto;
   margin-right: auto;
   color: white;
+}
+
+.blog-hero.category h1 {
+  margin-top: 0;
+}
+
+.blog-hero.category p {
+  margin: 0;
+}
+
+.blog-hero.category > div {
+  max-width: 1168px;
+  margin: 0 auto;
+}
+
+.blog-hero.category > div > div {
+  margin: 0 auto;
+  width: 80%;
+}
+
+.blog-hero {
+  position: relative;
+  padding: 32px;
+  min-height: 300px;
 }
 
 .blog-hero picture {
@@ -41,7 +70,7 @@ main .blog-hero-container {
 }
 
 @media (width >= 768px) {
-  .blog-hero > div > div {
+  .blog-hero.category > div > div {
     width: 33%;
   }
 }
